requirements.txt
tortoise-orm[accel]
python-dotenv
pycryptodome
pyrogram
tgcrypto
pydantic
loguru
pbkdf2
